--- a/ubc/ubc_noasm.go
+++ b/ubc/ubc_noasm.go
@@ -3,10 +3,12 @@
 
 package ubc
 
-// Check takes as input an expanded message block and verifies the unavoidable bitconditions
-// for all listed DVs. It returns a dvmask where each bit belonging to a DV is set if all
-// unavoidable bitconditions for that DV have been met.
-// Thus, one needs to do the recompression check for each DV that has its bit set.
+// CalculateDvMask takes as input an expanded message block and
+// verifies the unavoidable bitconditions for all listed DVs. It returns
+// a dvmask where each bit belonging to a DV is set if all unavoidable
+// bitconditions for that DV have been met.
+//
+//go:nosplit
 func CalculateDvMask(W [80]uint32) uint32 {
 	return CalculateDvMaskGeneric(W)
 }